% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workpath_data.R
\name{get_workpath_data}
\alias{get_workpath_data}
\title{Use workpath api to query data}
\usage{
get_workpath_data(
  your_company_subdomain,
  endpoint,
  auth_header = workpath_api_token(),
  raw_data = FALSE
)
}
\arguments{
\item{your_company_subdomain}{string: Company subdomain: https://<your-company>.workpath.com/api/v1}

\item{endpoint}{string:  Endpoint}

\item{auth_header}{string: Workpath API token}

\item{raw_data}{boolean: Raw data to return or not}
}
\value{
dataframe
}
\description{
Use workpath api to query data
}
