% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workpath_data.R
\name{raw_to_dataframe}
\alias{raw_to_dataframe}
\title{Return a dataframe object}
\usage{
raw_to_dataframe(resp)
}
\arguments{
\item{resp}{jsonlite::fromJSON object}
}
\value{
dataframe
}
\description{
Return a dataframe object
}
