% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{workpath_api_token}
\alias{workpath_api_token}
\title{Load api token if present. If not append token to .Renviron}
\usage{
workpath_api_token(
  your_company_subdomain,
  token,
  default_path_to_token = "~/.Renviron"
)
}
\arguments{
\item{your_company_subdomain}{Character sring api key}

\item{token}{Character sring api key}

\item{default_path_to_token}{Character sring api key}
}
\value{
auth_header
}
\description{
Load api token if present. If not append token to .Renviron
}
